+++ src/ptpapi/scripts/ptp.py
@@ -171,12 +171,12 @@ def search_page(api, args, target, movies, torrents, terms):
     if not movies and not torrents:
         logger.debug('Attempting to search target "%s" with terms %s', target, terms)
         if target == "torrents":
-            if terms.get("type") == "snatched":
+            if terms.get("type") in ["snatched", "uploaded", "seeding", "leeching"]:
                 if "userid" in terms:
                     user = ptpapi.User(terms["userid"])
                 else:
                     user = api.current_user()
-                movies = user.snatched(terms)
+                movies = user.search(terms.get("type"), terms)
             else:
                 movies = api.search(terms)
             # Check to see if we should scrape the cover view data to save calls