+++ src/ptpapi/user.py
@@ -21,10 +21,10 @@ class User:
     def __str__(self):
         return "<ptpapi.User ID %s>" % self.ID
 
-    def snatched(self, filters=None):
+    def search(self, search_type, filters=None):
         if filters is None:
             filters = {}
-        filters["type"] = "snatched"
+        filters["type"] = search_type
         filters["userid"] = str(self.ID)
         req = session.base_get("torrents.php", params=filters)
         return [Movie(data=m) for m in snarf_cover_view_data(req.content)]